+++ build.gradle
@@ -75,7 +75,7 @@ jacocoTestCoverageVerification {
 			limit {
 				counter = 'BRANCH'
 				value = 'COVEREDRATIO'
-				minimum = 0.60
+				minimum = 0.50
 			}
 		}
 	}
