+++ src/main/java/com/nodaji/payment/service/AccountServiceImpl.java
@@ -9,12 +9,10 @@ import com.nodaji.payment.global.domain.dto.WinDepositDto;
 import com.nodaji.payment.dto.request.WithdrawRequestDto;
 import com.nodaji.payment.dto.response.BuyResponseDto;
 import com.nodaji.payment.dto.response.PointResponseDto;
-import com.nodaji.payment.global.concurrency.utils.DistributedLock;
+import com.nodaji.payment.utils.DistributedLock;
 import com.nodaji.payment.global.domain.entity.Account;
-import com.nodaji.payment.global.domain.entity.History;
 import com.nodaji.payment.global.domain.exception.*;
 import com.nodaji.payment.global.domain.repository.AccountRepository;
-import com.nodaji.payment.global.domain.repository.HistoryRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.kafka.annotation.KafkaListener;
